Update Atividade00
Change variable type

+++ Atividade00.hs
@@ -1,6 +1,6 @@
 import Text.Printf
 
-addA :: Double -> Double
+addA :: Int -> Int
 addA a = a + 1
 
 main :: IO ()
@@ -8,5 +8,5 @@ main = do
    
     inputA <- getLine
     
-    let a = read inputA :: Double
+    let a = read inputA :: Int
     print (addA a)
